feat: combine with smithy-client
