+++ packages/smithy-client/src/index.ts
@@ -1,2 +1,5 @@
 export * from "./client";
 export * from "./command";
+export * from "./document-type";
+export * from "./exception";
+export * from "./isa";
