--- packages/smithy-core/.gitignore
@@ -1,7 +1,0 @@
-/node_modules/
-/build/
-/coverage/
-/docs/
-*.tgz
-*.log
-package-lock.json
