--- packages/smithy-core/.npmignore
@@ -1,12 +1,0 @@
-/src/
-/coverage/
-/docs/
-tsconfig.test.json
-
-*.spec.js
-*.spec.d.ts
-*.spec.js.map
-
-*.fixture.js
-*.fixture.d.ts
-*.fixture.js.map
