--- packages/smithy-core/README.md
@@ -1,4 +1,0 @@
-# @aws-sdk/@aws-sdk/smithy-core
-
-[![NPM version](https://img.shields.io/npm/v/@aws-sdk/@aws-sdk/smithy-core/preview.svg)](https://www.npmjs.com/package/@aws-sdk/@aws-sdk/smithy-core)
-[![NPM downloads](https://img.shields.io/npm/dm/@aws-sdk/@aws-sdk/smithy-core.svg)](https://www.npmjs.com/package/@aws-sdk/@aws-sdk/smithy-core)
