--- packages/smithy-core/package.json
@@ -1,26 +1,0 @@
-{
-  "name": "@aws-sdk/smithy-core",
-  "version": "0.1.0-preview.1",
-  "scripts": {
-    "prepublishOnly": "tsc",
-    "pretest": "tsc -p tsconfig.test.json",
-    "test": "jest --passWithNoTests"
-  },
-  "main": "./build/index.js",
-  "types": "./build/index.d.ts",
-  "author": {
-    "name": "AWS SDK for JavaScript Team",
-    "url": "https://aws.amazon.com/javascript/"
-  },
-  "license": "Apache-2.0",
-  "dependencies": {
-    "@aws-sdk/types": "^0.1.0-preview.5",
-    "@aws-sdk/middleware-stack": "^0.1.0-preview.6",
-    "tslib": "^1.8.0"
-  },
-  "devDependencies": {
-    "@types/jest": "^24.0.12",
-    "typescript": "~3.4.0",
-    "jest": "^24.7.1"
-  }
-}
